adminserver/src/requirements.txt
mysql-connector
requests
flask